backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.10
pandas>=2.2
psycopg2-binary>=2.9
python-dotenv>=1.0
holidays>=0.93
pytest>=8.3
httpx>=0.28
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=3.1,<5.0  # passlib 1.7.x incompatible with bcrypt>=5 (API change)
